api/ShelfScan-main/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-extensions==3.2.3
djangorestframework==3.15.1
phonenumbers==8.13.34
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.4.4
tzdata==2024.1
